+++ MediathekArr/Models/HistoryWrapper.cs
@@ -1,0 +1,9 @@
+﻿using System.Text.Json.Serialization;
+
+namespace MediathekArr.Models;
+
+public class HistoryWrapper
+{
+    [JsonPropertyName("history")]
+    public SabnzbdHistory History { get; set; }
+}
