+++ MediathekArr/Models/QueueWrapper.cs
@@ -1,0 +1,9 @@
+﻿using System.Text.Json.Serialization;
+
+namespace MediathekArr.Models;
+
+public class QueueWrapper
+{
+    [JsonPropertyName("queue")]
+    public SabnzbdQueue Queue { get; set; }
+}
