+++ MediathekArrLib/Models/Channel.cs
@@ -1,0 +1,19 @@
+﻿using System.Xml;
+using System.Xml.Serialization;
+
+namespace MediathekArrLib.Models;
+
+public class Channel
+{
+    [XmlElement("title")]
+    public string Title { get; set; }
+
+    [XmlElement("description")]
+    public string Description { get; set; }
+
+    [XmlElement("newznab:response", Namespace = "http://www.newznab.com/DTD/2010/feeds/attributes/")]
+    public NewznabResponse Response { get; set; }
+
+    [XmlElement("item")]
+    public List<Item> Items { get; set; } = [];
+}
