+++ MediathekArrLib/Models/Enclosure.cs
@@ -1,0 +1,16 @@
+﻿using System.Xml;
+using System.Xml.Serialization;
+
+namespace MediathekArrLib.Models;
+
+public class Enclosure
+{
+    [XmlAttribute("url")]
+    public string Url { get; set; }
+
+    [XmlAttribute("length")]
+    public long Length { get; set; }
+
+    [XmlAttribute("type")]
+    public string Type { get; set; }
+}
