+++ MediathekArrLib/Models/Rulesets/IdentificationResult.cs
@@ -1,4 +1,3 @@
-﻿namespace MediathekArrLib.Models.Rulesets
-{
-    public record IdentificationResult(string UsedRuleset, string Name, string GermanName, int? SeasonNumber, int? EpisodeNumber, string ItemTitle, TvdbEpisode MatchedEpisode);
-}
+﻿namespace MediathekArrLib.Models.Rulesets;
+
+public record IdentificationResult(string UsedRuleset, string Name, string GermanName, int? SeasonNumber, int? EpisodeNumber, string ItemTitle, TvdbEpisode MatchedEpisode);
