+++ MediathekArrLib/Models/Rulesets/MatchType.cs
@@ -1,0 +1,10 @@
+﻿namespace MediathekArrLib.Models.Rulesets;
+
+public enum MatchType
+{
+    ExactMatch,
+    Contains,
+    Regex,
+    GreaterThan,
+    LessThan
+}
