+++ MediathekArrLib/Models/Rulesets/MatchedEpisodeInfo.cs
@@ -1,4 +1,3 @@
-﻿namespace MediathekArrLib.Models.Rulesets
-{
-    public record MatchedEpisodeInfo(TvdbEpisode Episode, ApiResultItem Item, string ShowName, string MatchedTitle);
-}
+﻿namespace MediathekArrLib.Models.Rulesets;
+
+public record MatchedEpisodeInfo(TvdbEpisode Episode, ApiResultItem Item, string ShowName, string MatchedTitle);
