+++ MediathekArrLib/Models/Rulesets/MatchingStrategy.cs
@@ -1,12 +1,9 @@
-﻿using System.Text.Json.Serialization;
+﻿namespace MediathekArrLib.Models.Rulesets;
 
-namespace MediathekArrLib.Models.Rulesets
+public enum MatchingStrategy
 {
-    public enum MatchingStrategy
-    {
-        SeasonAndEpisodeNumber, // Use season + episode number for matching
-        ItemTitleIncludes,      // Match episodes where the tvdb episode name contains this title
-        ItemTitleExact,          // Match episodes with an exact itemTitle
-        ItemTitleEqualsAirdate
-    }
+    SeasonAndEpisodeNumber, // Use season + episode number for matching
+    ItemTitleIncludes,      // Match episodes where the tvdb episode name contains this title
+    ItemTitleExact,          // Match episodes with an exact itemTitle
+    ItemTitleEqualsAirdate
 }
