+++ MediathekArrLib/Models/Rulesets/Pagination.cs
@@ -1,0 +1,18 @@
+﻿using System.Text.Json.Serialization;
+
+namespace MediathekArrLib.Models.Rulesets;
+
+public class Pagination
+{
+    [JsonPropertyName("currentPage")]
+    public int CurrentPage { get; set; }
+
+    [JsonPropertyName("totalPages")]
+    public int TotalPages { get; set; }
+
+    [JsonPropertyName("totalItems")]
+    public int TotalItems { get; set; }
+
+    [JsonPropertyName("itemsPerPage")]
+    public int ItemsPerPage { get; set; }
+}
