+++ MediathekArrLib/Models/Rulesets/RegexRule.cs
@@ -1,13 +1,12 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace MediathekArrLib.Models.Rulesets
-{
-    public class RegexRule
-    {
-        [JsonPropertyName("field")]
-        public string Field { get; set; } = string.Empty;
+namespace MediathekArrLib.Models.Rulesets;
 
-        [JsonPropertyName("pattern")]
-        public string Pattern { get; set; } = string.Empty;
-    }
+public class RegexRule
+{
+    [JsonPropertyName("field")]
+    public string Field { get; set; } = string.Empty;
+
+    [JsonPropertyName("pattern")]
+    public string Pattern { get; set; } = string.Empty;
 }
