+++ MediathekArrLib/Models/Rulesets/TitleRegexRuleType.cs
@@ -1,0 +1,7 @@
+﻿namespace MediathekArrLib.Models.Rulesets;
+
+public enum TitleRegexRuleType
+{
+    Static, // Static text to include in the title
+    Regex   // Regex to extract text from an API field
+}
