+++ MediathekArrLib/Models/TvdbAlias.cs
@@ -1,0 +1,3 @@
+﻿namespace MediathekArrLib.Models;
+
+public record TvdbAlias(string Language, string Name);
